Improvement: 프로필 컴포넌트화
